Move serial_task into celery app
