+++ corehq/form_processor/tasks.py
@@ -9,10 +9,9 @@ from couchforms.models import UnfinishedSubmissionStub
 from dimagi.utils.couch import CriticalSection
 from dimagi.utils.logging import notify_exception
 
-from corehq.apps.celery import periodic_task
+from corehq.apps.celery import periodic_task, serial_task
 from corehq.form_processor.reprocess import reprocess_unfinished_stub
 from corehq.util.celery_utils import no_result_task
-from corehq.util.decorators import serial_task
 from corehq.util.metrics import metrics_counter, metrics_gauge
 from corehq.util.metrics.const import MPM_MAX
